+++ src/mm_write.h
@@ -7,6 +7,6 @@
  * cooperative multitasking engine.
 */
 
-int mm_write(mm_io_t*, char*, int, uint64_t);
+int mm_write(mm_io_t*, char*, int, uint32_t);
 
 #endif /* MM_WRITE_H */
